requirements.txt
streamlit
groq
python-dotenv
pydantic
typing
datetime
